session

+++ app/api/calendar/[id]/route.ts
@@ -1,6 +1,8 @@
 // app/api/calendar/[id]/route.ts
 import { NextResponse } from 'next/server';
 import { CalendarEventService } from "../service";
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../../auth/[...nextauth]/route';
 
 const eventService = new CalendarEventService();
 
@@ -8,6 +10,12 @@ export async function GET(
   request: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
+
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.redirect(new URL('/unauthorized', request.url));  
+  }
+
   try {
     const { id } = await params;
     const event = await eventService.getEventById(Number(id));
@@ -32,6 +40,11 @@ export async function PUT(
   request: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.redirect(new URL('/unauthorized', request.url));  
+  }
+
   try {
     const { id } = await params;
     const event = await request.json();
@@ -49,6 +62,12 @@ export async function DELETE(
   request: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
+
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.redirect(new URL('/unauthorized', request.url));  
+  }
+  
   try {
     const { id } = await params;
     await eventService.deleteEvent(Number(id));
